+++ src/Input.h
@@ -15,7 +15,7 @@ protected:
 	virtual bool IsMouseButtonPressedImpl(int button) = 0;
 	virtual float GetMouseXImpl() = 0;
 	virtual float GetMouseYImpl() = 0;
-	virtual std::pair<float, float> GetMousePositionImpl() = 0;
+	virtual std::pair<double, double> GetMousePositionImpl() = 0;
 private:
 	static Input* s_Instance;
 };
